updated the graphs and added it to labs component
