+++ src/lab/Graphs/AreaGraph/AreaGraph.stories.tsx
@@ -1,18 +1,8 @@
-// import generateDateValue, {
-//   DateValue,
-// } from "@visx/mock-data/lib/generators/genDateValue";
+import { storiesOf } from '@storybook/react';
 import React from 'react';
-import BrushChart from './AreaGraph-Scroller';
-// import ParentSize from "@visx/responsive/lib/components/ParentSize";
-
-export interface DataValue {
-  date: number;
-  value: number;
-}
-export interface AreaGrapher {
-  metricName: string;
-  data: Array<DataValue>;
-}
+import { ThemedBackground } from '../../../utils/storybook';
+import { AreaGraph } from './AreaGraph';
+import { AreaGrapher, DataValue } from './base';
 
 const data1: DataValue[] = [
   { date: 1000, value: 20 },
@@ -35,32 +25,21 @@ const data2: DataValue[] = [
   { date: 8000, value: 10 },
 ];
 
-// const w:AreaGrapher={date:,value:20}
 const seriestest: Array<AreaGrapher> = [
   { metricName: 'chaos-exporter', data: data1 },
 ];
 const seriestest2: Array<AreaGrapher> = [{ metricName: 'heptio', data: data2 }];
 
-// let lineCount: number = 2;
-
-// console.log(seriestest);
-
-// const width = 600;
-// const height = 400;
-// console.log(seriestest);
-function BaseArea() {
-  return (
-    <div className="App">
-      <BrushChart
+storiesOf('Graphs/Area', module)
+  // Litmus Portal
+  .add('Kubera Chaos', () => (
+    <ThemedBackground platform="kubera-chaos">
+      <AreaGraph
         width={600}
         height={400}
         closedSeries={seriestest}
         openSeries={seriestest2}
         showPoints={true}
       />
-      ,
-    </div>
-  );
-}
-
-export default BaseArea;
+    </ThemedBackground>
+  ));
