+++ src/lab/Graphs/AreaGraph/base.ts
@@ -1,0 +1,9 @@
+export interface DataValue {
+  date: number;
+  value: number;
+}
+
+export interface AreaGrapher {
+  metricName: string;
+  data: Array<DataValue>;
+}
