+++ src/lab/Graphs/AreaGraph/index.ts
@@ -1,0 +1,1 @@
+export * from './AreaGraph';
